+++ example/main.dart
@@ -1,30 +1,40 @@
-import 'package:rest_api_impl/data/data_sources/i_data_sources/i_image_service_rest_api_data_source.dart';
-import 'package:rest_api_impl/rest_api_impl.dart';
-import 'package:get_it/get_it.dart';
 import 'dart:io';
 
-final sl = GetIt.instance;
+import 'package:cross_file/cross_file.dart';
+import 'package:flutter/material.dart';
+import 'package:rest_api_impl/rest_api_impl.dart';
 
 void main() async {
+  WidgetsFlutterBinding.ensureInitialized();
+
+  registerRestApiDataSourceServiceGetItDI(
+    baseUrl: RestApiConfigConst.BASE_URL,
+    apiKey: RestApiConfigConst.API_KEY,
+    clientType: RestApiClientType.dio,
+  );
+
   final restApiService = sl<IRestApiCrudService>();
   final imageService = sl<IImageServiceRestApiDataSource>();
 
-  // Example: Create User
   final response = await restApiService.addData(
     endPoint: 'user/create',
     data: {'name': 'Alice', 'email': 'alice@example.com'},
   );
-  print('Create User Response: $response');
+  debugPrint('Create User Response: $response');
 
-  // Example: Get User Data
   final userData = await restApiService.getData(endPoint: 'user/1');
   print('User Data: $userData');
 
-  // Example: Upload File
   final file = File('path/to/image.png');
+  final xFile = XFile(file.path);
   final uploadResponse = await imageService.uploadFile(
-    file: file,
+    file: xFile,
     endPoint: 'upload/image',
   );
   print('Upload Response: $uploadResponse');
+}
+
+class RestApiConfigConst {
+  static const BASE_URL = 'https://api.example.com/';
+  static const API_KEY = 'xyz-123-token';
 }