+++ lib/data/data_sources/image_service/i_image_service/i_image_service_rest_api_data_source.dart
@@ -1,8 +1,8 @@
-import 'dart:io';
+import 'package:cross_file/cross_file.dart';
 
 abstract class IImageServiceRestApiDataSource {
   Future<String?> uploadFile({
-    required File file,
+    required XFile file,
     required String endPoint,
     Map<String, dynamic>? queryParams,
     String? fileName,