+++ lib/enums/rest_api_client_type.dart
@@ -1,0 +1,4 @@
+enum RestApiClientType {
+  dio,
+  http,
+}